requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.7
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
openpyxl==3.1.1
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
psycopg2-binary==2.9.5
pydantic==1.10.5
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
yarl==1.8.2
